# backend/api/gen/go/admin/service/v1/i_post_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_post.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-cms/api/gen/go/content/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPostServiceCreate = "/admin.service.v1.PostService/Create"
const OperationPostServiceDelete = "/admin.service.v1.PostService/Delete"
const OperationPostServiceGet = "/admin.service.v1.PostService/Get"
const OperationPostServiceList = "/admin.service.v1.PostService/List"
const OperationPostServiceTranslationExists = "/admin.service.v1.PostService/TranslationExists"
const OperationPostServiceUpdate = "/admin.service.v1.PostService/Update"

type PostServiceHTTPServer interface {
	// Create 创建帖子
	Create(context.Context, *v11.CreatePostRequest) (*v11.Post, error)
	// Delete 删除帖子
	Delete(context.Context, *v11.DeletePostRequest) (*emptypb.Empty, error)
	// Get 获取帖子数据
	Get(context.Context, *v11.GetPostRequest) (*v11.Post, error)
	// List 获取帖子列表
	List(context.Context, *v1.PagingRequest) (*v11.ListPostResponse, error)
	// TranslationExists 检查翻译是否存在
	TranslationExists(context.Context, *v11.PostTranslationExistsRequest) (*v11.PostTranslationExistsResponse, error)
	// Update 更新帖子
	Update(context.Context, *v11.UpdatePostRequest) (*v11.Post, error)
}

func RegisterPostServiceHTTPServer(s *http.Server, srv PostServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/posts", _PostService_List24_HTTP_Handler(srv))
	r.GET("/admin/v1/posts/{id}", _PostService_Get24_HTTP_Handler(srv))
	r.POST("/admin/v1/posts", _PostService_Create18_HTTP_Handler(srv))
	r.PUT("/admin/v1/posts/{id}", _PostService_Update18_HTTP_Handler(srv))
	r.DELETE("/admin/v1/posts/{id}", _PostService_Delete18_HTTP_Handler(srv))
	r.GET("/admin/v1/posts/{post_id}/translations/{language_code}", _PostService_TranslationExists0_HTTP_Handler(srv))
}

func _PostService_List24_HTTP_Handler(srv PostServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPostServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPostResponse)
		return ctx.Result(200, reply)
	}
}

func _PostService_Get24_HTTP_Handler(srv PostServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetPostRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPostServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetPostRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Post)
		return ctx.Result(200, reply)
	}
}

func _PostService_Create18_HTTP_Handler(srv PostServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreatePostRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPostServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreatePostRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Post)
		return ctx.Result(200, reply)
	}
}

func _PostService_Update18_HTTP_Handler(srv PostServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdatePostRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPostServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdatePostRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Post)
		return ctx.Result(200, reply)
	}
}

func _PostService_Delete18_HTTP_Handler(srv PostServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeletePostRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPostServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeletePostRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _PostService_TranslationExists0_HTTP_Handler(srv PostServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.PostTranslationExistsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPostServiceTranslationExists)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TranslationExists(ctx, req.(*v11.PostTranslationExistsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.PostTranslationExistsResponse)
		return ctx.Result(200, reply)
	}
}

type PostServiceHTTPClient interface {
	// Create 创建帖子
	Create(ctx context.Context, req *v11.CreatePostRequest, opts ...http.CallOption) (rsp *v11.Post, err error)
	// Delete 删除帖子
	Delete(ctx context.Context, req *v11.DeletePostRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 获取帖子数据
	Get(ctx context.Context, req *v11.GetPostRequest, opts ...http.CallOption) (rsp *v11.Post, err error)
	// List 获取帖子列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListPostResponse, err error)
	// TranslationExists 检查翻译是否存在
	TranslationExists(ctx context.Context, req *v11.PostTranslationExistsRequest, opts ...http.CallOption) (rsp *v11.PostTranslationExistsResponse, err error)
	// Update 更新帖子
	Update(ctx context.Context, req *v11.UpdatePostRequest, opts ...http.CallOption) (rsp *v11.Post, err error)
}

type PostServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPostServiceHTTPClient(client *http.Client) PostServiceHTTPClient {
	return &PostServiceHTTPClientImpl{client}
}

// Create 创建帖子
func (c *PostServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreatePostRequest, opts ...http.CallOption) (*v11.Post, error) {
	var out v11.Post
	pattern := "/admin/v1/posts"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPostServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除帖子
func (c *PostServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeletePostRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/posts/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPostServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 获取帖子数据
func (c *PostServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetPostRequest, opts ...http.CallOption) (*v11.Post, error) {
	var out v11.Post
	pattern := "/admin/v1/posts/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPostServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 获取帖子列表
func (c *PostServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListPostResponse, error) {
	var out v11.ListPostResponse
	pattern := "/admin/v1/posts"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPostServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// TranslationExists 检查翻译是否存在
func (c *PostServiceHTTPClientImpl) TranslationExists(ctx context.Context, in *v11.PostTranslationExistsRequest, opts ...http.CallOption) (*v11.PostTranslationExistsResponse, error) {
	var out v11.PostTranslationExistsResponse
	pattern := "/admin/v1/posts/{post_id}/translations/{language_code}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPostServiceTranslationExists))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新帖子
func (c *PostServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdatePostRequest, opts ...http.CallOption) (*v11.Post, error) {
	var out v11.Post
	pattern := "/admin/v1/posts/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPostServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}