# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-cms
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-cms/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/posts:
        get:
            tags:
                - PostService
            description: 获取帖子列表
            operationId: PostService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPostResponse'
        post:
            tags:
                - PostService
            description: 创建帖子
            operationId: PostService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Post'
    /app/v1/posts/{id}:
        get:
            tags:
                - PostService
            description: 获取帖子数据
            operationId: PostService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: code
                  in: query
                  schema:
                    type: string
                - name: locale
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Post'
        put:
            tags:
                - PostService
            description: 更新帖子
            operationId: PostService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Post'
        delete:
            tags:
                - PostService
            description: 删除帖子
            operationId: PostService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreatePostRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Post'
            description: 请求 - 创建帖子
...
        ListPostResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Post'
                total:
                    type: string
            description: 回应 - 帖子列表
...
        Post:
            type: object
            properties:
                id:
                    type: integer
                    description: 帖子ID
                    format: uint32
                status:
                    enum:
                        - POST_STATUS_UNSPECIFIED
                        - POST_STATUS_DRAFT
                        - POST_STATUS_PUBLISHED
                        - POST_STATUS_SCHEDULED
                        - POST_STATUS_TRASHED
                    type: string
                    description: 帖子状态
                    format: enum
                editorType:
                    enum:
                        - EDITOR_TYPE_UNSPECIFIED
                        - EDITOR_TYPE_MARKDOWN
                        - EDITOR_TYPE_RICH_TEXT
                        - EDITOR_TYPE_PLAIN_TEXT
                        - EDITOR_TYPE_CODE
                        - EDITOR_TYPE_JSON_BLOCK
                        - EDITOR_TYPE_VISUAL_BUILDER
                    type: string
                    description: 编辑器类型
                    format: enum
                code:
                    type: string
                    description: 唯一编码（必须唯一，建议使用英文）
                disallowComment:
                    type: boolean
                    description: 帖子是否禁止评论
                inProgress:
                    type: boolean
                    description: 帖子是否正在编辑中（如正在编辑则不允许发布，避免内容不完整）
                autoSummary:
                    type: boolean
                    description: 是否自动生成摘要
                isFeatured:
                    type: boolean
                    description: 是否推荐
                sortOrder:
                    type: integer
                    description: 排序优先级（数值越小越靠前，同组内排序）
                    format: uint32
                visits:
                    type: integer
                    description: 帖子访问次数
                    format: int32
                likes:
                    type: integer
                    description: 帖子点赞次数
                    format: int32
                commentCount:
                    type: integer
                    description: 帖子评论数
                    format: int32
                authorId:
                    type: integer
                    description: 评论作者ID，0表示游客
                    format: uint32
                authorName:
                    type: string
                    description: 帖子作者名称（游客填写）
                customFields:
                    type: object
                    additionalProperties:
                        type: string
                    description: 自定义字段，键值对形式，便于扩展
                translations:
                    type: array
                    items:
                        $ref: '#/components/schemas/PostTranslation'
                    description: 多语言翻译列表
                availableLanguages:
                    example: ["zh-CN", "en-US", "ja-JP"]
                    type: array
                    items:
                        type: string
                    description: 可用的语言代码列表（快速查询，避免遍历 translations）
                categoryIds:
                    type: array
                    items:
                        type: integer
                        format: uint32
                    description: 关联的分类ID列表（多选）
                tagIds:
                    type: array
                    items:
                        type: integer
                        format: uint32
                    description: 关联的标签ID列表（多选）
                passwordHash:
                    type: string
                    description: 密码哈希
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
                publishTime:
                    type: string
                    description: 发布时间
                    format: date-time
            description: 帖子
        PostTranslation:
            type: object
            properties:
                id:
                    type: integer
                    description: 帖子翻译ID
                    format: uint32
                postId:
                    type: integer
                    description: 关联的帖子ID
                    format: uint32
                languageCode:
                    type: string
                    description: 语言代码
                title:
                    type: string
                    description: 帖子标题
                slug:
                    type: string
                    description: 语言特定 slug（覆盖主表 slug，如中文用拼音 /post/hou-duan）
                summary:
                    type: string
                    description: 帖子摘要
                content:
                    type: string
                    description: 帖子内容
                originalContent:
                    type: string
                    description: 原始内容
                thumbnail:
                    type: string
                    description: 缩略图（某些语言可能需要不同图片）
                template:
                    type: string
                    description: 模板名称（某些语言可能使用不同模板）
                fullPath:
                    type: string
                    description: 完整路径（如 /zh-CN/blog/post-slug）
                wordCount:
                    type: integer
                    description: 当前语言版本的字数（中文按字符数，英文按单词数）
                    format: uint32
                metaKeywords:
                    type: string
                    description: SEO 关键词
                metaDescription:
                    type: string
                    description: SEO 描述
                seoTitle:
                    type: string
                    description: SEO 标题
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 帖子翻译
...
        UpdatePostRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/Post'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 请求 - 更新帖子
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...